refactor: 统一缓存配置并重构缓存实现
- 将默认TTL配置移至顶层cache配置
- 更新moka和redis缓存实现使用统一TTL配置
- 修改缓存traits中的ttl参数类型为u64
- 清理各配置文件中重复的TTL配置项

feat: 添加班级学生管理功能

- 实现列出班级学生接口
- 添加更新学生信息接口
- 完善相关请求和响应数据结构

refactor: 重命名storages模块为repository

- 更新所有相关引用路径
- 保持接口和功能不变

perf: 优化JWT中间件性能

- 添加用户信息缓存
- 减少数据库查询次数
- 添加详细的调试日志

+++ src/cache/traits.rs
@@ -20,7 +20,7 @@ pub trait ObjectCache: Send + Sync {
     async fn get_raw(&self, key: &str) -> CacheResult<String>;
 
     /// 插入原始 JSON 字符串
-    async fn insert_raw(&self, key: String, value: String, ttl: u32);
+    async fn insert_raw(&self, key: String, value: String, ttl: u64);
 
     /// 移除指定键
     async fn remove(&self, key: &str);
@@ -41,7 +41,7 @@ pub trait TypedObjectCache {
         &self,
         key: String,
         value: T,
-        ttl: u32,
+        ttl: u64,
     ) -> impl std::future::Future<Output = ()> + Send
     where
         T: Serialize + Send;
@@ -87,7 +87,7 @@ impl<C: ObjectCache + ?Sized> TypedObjectCache for C {
         }
     }
 
-    async fn insert<T>(&self, key: String, value: T, ttl: u32) -> ()
+    async fn insert<T>(&self, key: String, value: T, ttl: u64) -> ()
     where
         T: Serialize + Send,
     {
